Cleaning up the code, spliting place pixel into individual files
