+++ src/graphics/font/spk_font_atlas.cpp
@@ -2,10 +2,24 @@
 
 namespace spk
 {
+	void _applyOutline(std::vector<uint8_t> &p_atlasData, const spk::Vector2Int &p_atlasSize, const spk::Font::Key &p_key);
+
 	Font::Atlas::Atlas()
 	{
 	}
 
+	void _normalizeAtlasData(std::vector<uint8_t> &p_atlasData, const spk::Vector2Int &p_atlasSize)
+	{
+		size_t maxIndex = static_cast<size_t>(p_atlasSize.x * p_atlasSize.y);
+		for (size_t i = 0; i < maxIndex; i++)
+		{
+			if (p_atlasData[i] != 0xFF)
+				p_atlasData[i] = 0x00;
+			else
+				p_atlasData[i] = 0xFF;
+		}
+	}
+	
 	Font::Atlas::Atlas(const std::vector<uint8_t> &p_fontData, const Configuration &p_fontConfiguration, const Key &p_key) :
 		_glyphDatas(p_fontConfiguration.nbGlyph())
 	{
